pyproject.toml
[project]
name = "django-docs-helper"
version = "0.1.0"
description = ""
authors = [
    {name = "Freddy Hidalgo-Monchez",email = "freddy.hidalgo.m@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4"
dependencies = [
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "langgraph (>=0.2.63,<0.3.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-community (>=0.3.14,<0.4.0)",
    "tavily-python (>=0.5.0,<0.6.0)",
    "langchain-chroma (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "black (>=24.10.0,<25.0.0)",
    "isort (>=5.13.2,<6.0.0)",
    "langchain-openai (==0.2.12)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
